+++ lib/services/service.go
@@ -6,6 +6,11 @@
 package services
 
 /*
+The identifier used for services with the network.
+*/
+type ServiceID string
+
+/*
 The developer should implement their own use-case-specific interface that extends this one
  */
 type Service interface {